# yaml/swagger.yaml
swagger: '2.0'
info:
  description: API de la pràctica de l'assignatura ASW de la FIB sobre Hacker News.
  version: 1.0.0
  title: Swagger Hacker News
host: 'practicaasw-hackernews.herokuapp.com'
basePath: /api/v1
schemes:
  - https
produces:
  - application/json
securityDefinitions:
  ApiKeyAuth:
    type: apiKey
    in: header
    name: Authorization
    description: És la clau d'usuari generada per l'aplicació de Google.
paths:
...
  /submissions:
    get:
      tags:
        - submissions
      summary: Obté totes les submissions
      description: Obté totes les submissions que hi ha a Hacker News
      responses:
        200:
          description: successful operation
          schema:
            type: array
            items:
              $ref: '#/definitions/SubmissionObject'
        default:
          description: Error inesperat.
          schema:
            $ref: '#/definitions/Error'
            
    post:
      security:
        - ApiKeyAuth: []
      summary: Crea una nova Submission
      description: Es crea una nova submission a Hacker News
      tags:
        - submissions
      parameters:
        - name: SubmissionBody
          in: body
          schema:
            $ref: '#/definitions/SubmissionBody'
      responses:
        200:
          description: successful operation
          schema:
            type: array
            items:
              $ref: '#/definitions/SubmissionObject'
        403:
          description: Error in authenticity.
          schema:
            $ref: '#/definitions/Error'
        417:
          description: Error in parameters.
          schema:
            $ref: '#/definitions/Error'
        500:
          description: Submission not saved.
          schema:
            $ref: '#/definitions/Error'
        default:
          description: Error inesperat.
          schema:
            $ref: '#/definitions/Error'
...
  /replies/{id}/replies:
    get:
      tags:
        - replies
      summary: Obté les replies de la reply
      description: Obté les replies de la reply corresponents que hi ha a Hacker News
      parameters:
        - name: id
          in: path
          type: number
          required: true
          description: Identificador de la reply
      responses:
        200:
          description: successful operation
          schema:
            type: array
            items:
              $ref: '#/definitions/RepliesObject'
        404:
          description: Reply does not exist
          schema:
            $ref: '#/definitions/Error'
...
definitions:
...
  SubmissionObject:
    type: object
    properties:
      status:
        type: string
      message:
        type: string
      data:
        $ref: '#/definitions/Submission'
...
  Submission:
    type: object
    properties:
      id:
        type: integer
      title:
        type: string
      url:
        type: string
      text:
        type: string
      user_id:
        type: integer
      created_at:
        type: string
      updated_at:
        type: string
  SubmissionBody:
    type: object
    properties:
      title:
        type: string
      url:
        type: string
      text:
        type: string
        
  # Comments Objects
...
  RepliesObject:
    type: object
    properties:
      status:
        type: string
      message:
        type: string
      data:
        $ref: '#/definitions/Reply'
  Reply:
    type: object
    properties:
      id:
        type: integer
      content:
        type: string
      comment_id:
        type: integer
      user_id:
        type: integer
      created_at:
        type: string
      updated_at:
        type: string
      reply_parent_id:
        type: integer
      submission_id:
        type: integer
...
  Error:
    type: object
    properties:
      status:
        type: string
      message:
        type: string
      data:
        $ref: '#/definitions/ErrorArray'
  ErrorArray:
    type: object